requirements.txt
# Custom Support Bot UI - Dependencies
# Inherits most dependencies from sql-support-bot

# Web Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Core LangChain/LangGraph (same as sql-support-bot)
langchain-community>=0.3.0
langchain-core>=0.3.0
langgraph>=0.2.0
langchain-openai

# Database
sqlalchemy>=2.0.0

# Utils
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0

# External APIs
twilio>=9.0.0
google-api-python-client>=2.100.0
